# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MerchantAdmissionDto": {
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "application_status": {
            "title": "Application Status",
            "type": "string"
          },
          "bay_count": {
            "title": "Bay Count",
            "type": "integer"
          },
          "bay_photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bay Photo Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "ekyc_status": {
            "title": "Ekyc Status",
            "type": "string"
          },
          "go_live_blockers": {
            "items": {
              "type": "string"
            },
            "title": "Go Live Blockers",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ops_rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ops Rejection Reason"
          },
          "ops_reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ops Reviewed At"
          },
          "ops_reviewed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ops Reviewed By"
          },
          "payment_recipient_status": {
            "title": "Payment Recipient Status",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "photo_status": {
            "title": "Photo Status",
            "type": "string"
          },
          "pipeline_status": {
            "title": "Pipeline Status",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "storefront_photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storefront Photo Url"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "user_id",
          "name",
          "address",
          "latitude",
          "longitude",
          "bay_count",
          "status",
          "pipeline_status",
          "application_status",
          "photo_status",
          "payment_recipient_status",
          "ekyc_status",
          "created_at"
        ],
        "title": "MerchantAdmissionDto",
        "type": "object"
      },
      "MerchantApplicationRequest": {
        "properties": {
          "address": {
            "maxLength": 500,
            "minLength": 4,
            "title": "Address",
            "type": "string"
          },
          "bay_count": {
            "default": 1,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Bay Count",
            "type": "integer"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "operating_hours_end": {
            "default": "20:00",
            "maxLength": 16,
            "title": "Operating Hours End",
            "type": "string"
          },
          "operating_hours_start": {
            "default": "08:00",
            "maxLength": 16,
            "title": "Operating Hours Start",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "name",
          "address",
          "latitude",
          "longitude"
        ],
        "title": "MerchantApplicationRequest",
        "type": "object"
      },
...
      "MerchantEkycStatusResponse": {
        "properties": {
          "ekyc_status": {
            "title": "Ekyc Status",
            "type": "string"
          },
          "merchant_id": {
            "format": "uuid",
            "title": "Merchant Id",
            "type": "string"
          },
          "submissions": {
            "items": {
              "$ref": "#/components/schemas/MerchantEkycSubmissionDto"
            },
            "title": "Submissions",
            "type": "array"
          }
        },
        "required": [
          "merchant_id",
          "ekyc_status",
          "submissions"
        ],
        "title": "MerchantEkycStatusResponse",
        "type": "object"
      },
      "MerchantEkycSubmissionDto": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "object_key",
          "status",
          "submitted_at"
        ],
        "title": "MerchantEkycSubmissionDto",
        "type": "object"
      },
      "MerchantEkycSubmitRequest": {
        "properties": {
          "object_key": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Object Key",
            "type": "string"
          }
        },
        "required": [
          "object_key"
        ],
        "title": "MerchantEkycSubmitRequest",
        "type": "object"
      },
...
      "MerchantPaymentSetupRequest": {
        "properties": {
          "account_holder_name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Account Holder Name",
            "type": "string"
          },
          "account_number": {
            "maxLength": 80,
            "minLength": 4,
            "title": "Account Number",
            "type": "string"
          },
          "bank_name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Bank Name",
            "type": "string"
          },
          "qr_object_key": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qr Object Key"
          }
        },
        "required": [
          "bank_name",
          "account_number",
          "account_holder_name"
        ],
        "title": "MerchantPaymentSetupRequest",
        "type": "object"
      },
      "MerchantPhotoConfirmRequest": {
        "properties": {
          "bay_object_key": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Bay Object Key",
            "type": "string"
          },
          "storefront_object_key": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Storefront Object Key",
            "type": "string"
          }
        },
        "required": [
          "storefront_object_key",
          "bay_object_key"
        ],
        "title": "MerchantPhotoConfirmRequest",
        "type": "object"
      },
...
      "OpsMerchantDecisionRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "OpsMerchantDecisionRequest",
        "type": "object"
      },
...
      "PendingMerchantsResponse": {
        "properties": {
          "merchants": {
            "items": {
              "$ref": "#/components/schemas/MerchantAdmissionDto"
            },
            "title": "Merchants",
            "type": "array"
          }
        },
        "required": [
          "merchants"
        ],
        "title": "PendingMerchantsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Foundation OpenAPI contract for the TrueCare Python port.",
    "title": "TrueCare API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/merchants/applications": {
      "post": {
        "operationId": "post_v1_merchants_applications",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantAdmissionDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Application",
        "tags": [
          "merchant-admission",
          "merchant"
        ]
      }
    },
...
    "/v1/merchants/{id}/confirm-photo": {
      "post": {
        "operationId": "post_v1_merchants_by_id_confirm_photo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantPhotoConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantAdmissionDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Photo",
        "tags": [
          "merchant-admission",
          "merchant"
        ]
      }
    },
...
    "/v1/merchants/{id}/ekyc/bank": {
      "post": {
        "operationId": "post_v1_merchants_by_id_ekyc_bank",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantEkycSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantEkycStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ekyc Bank",
        "tags": [
          "merchant-admission",
          "merchant-ekyc"
        ]
      }
    },
    "/v1/merchants/{id}/ekyc/cmnd": {
      "post": {
        "operationId": "post_v1_merchants_by_id_ekyc_cmnd",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantEkycSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantEkycStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ekyc Cmnd",
        "tags": [
          "merchant-admission",
          "merchant-ekyc"
        ]
      }
    },
    "/v1/merchants/{id}/ekyc/selfie": {
      "post": {
        "operationId": "post_v1_merchants_by_id_ekyc_selfie",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantEkycSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantEkycStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ekyc Selfie",
        "tags": [
          "merchant-admission",
          "merchant-ekyc"
        ]
      }
    },
    "/v1/merchants/{id}/ekyc/status": {
      "get": {
        "operationId": "get_v1_merchants_by_id_ekyc_status",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantEkycStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ekyc Status",
        "tags": [
          "merchant-admission",
          "merchant-ekyc"
        ]
      }
    },
...
    "/v1/merchants/{id}/payment-setup": {
      "post": {
        "operationId": "post_v1_merchants_by_id_payment_setup",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantPaymentSetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantAdmissionDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Payment Setup",
        "tags": [
          "merchant-admission",
          "merchant"
        ]
      }
    },
...
    "/v1/ops/merchants/pending": {
      "get": {
        "operationId": "get_v1_ops_merchants_pending",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingMerchantsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Pending Merchants",
        "tags": [
          "merchant-admission",
          "ops"
        ]
      }
    },
    "/v1/ops/merchants/{id}/approve": {
      "post": {
        "operationId": "post_v1_ops_merchants_by_id_approve",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantAdmissionDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Approve Merchant",
        "tags": [
          "merchant-admission",
          "ops"
        ]
      }
    },
    "/v1/ops/merchants/{id}/reject": {
      "post": {
        "operationId": "post_v1_ops_merchants_by_id_reject",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpsMerchantDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantAdmissionDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Reject Merchant",
        "tags": [
          "merchant-admission",
          "ops"
        ]
      }
    },
    "/v1/ops/merchants/{id}/suspend": {
      "post": {
        "operationId": "post_v1_ops_merchants_by_id_suspend",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpsMerchantDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantAdmissionDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Suspend Merchant",
        "tags": [
          "merchant-admission",
          "ops"
        ]
      }
    },
    "/v1/ops/merchants/{id}/verify-payment-recipient": {
      "post": {
        "operationId": "post_v1_ops_merchants_by_id_verify_payment_recipient",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantAdmissionDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Verify Payment Recipient",
        "tags": [
          "merchant-admission",
          "ops"
        ]
      }
    },
...
  }
}